### 2019-12-21 #1980 @whedon
[PRE REVIEW]: NCBImeta: efficient and comprehensive metadata retrieval from NCBI databases
**Submitting author:** @ktmeaton (<a href="http://orcid.org/0000-0001-6862-7756">Katherine Eaton</a>)
**Repository:** <a href="https://github.com/ktmeaton/NCBImeta" target ="_blank">https://github.com/ktmeaton/NCBImeta</a>
**Version:** v0.6.1
