### 2019-12-21 #1980 @labarba
👋 @lpantano — I know you said you'd be on holiday for two weeks ... but so is most everyone on the JOSS board! This submission looks like a good fit for you to take through review. Can you take it, in the understanding it will take a little while to get started?
