### 2019-12-21 #1980 @ktmeaton
Thank you for considering this paper! From the spreadsheet of current JOSS reviewers, here are some suggestions based on topic (Bioinformatics/Genomics) and language (Python):
- Luiz Irber (luizirber)
- Louis J Taylor (louiejtaylor)
- Hadrien Gourlé (HadrienG)
- Joep de Ligt (jdeligt)
- Simone Coughlan (coughls)
- Rachel Schwartz (rachelss)
- Marek Borowiec (marekborowiec)
- C. Titus Brown  (ctb)
