### 2020-01-07 55e8bf7
Adding Bioconda as an installation option
