### 2020-01-07 c3fc9a4
Paper remove unnecessary reference ncbi website [skip ci]
