+++ paper/paper.md
@@ -39,7 +39,7 @@ In response, ``NCBImeta`` aims to provide a more user-inclusive experience to me
 
 # NCBImeta
 
-``NBCImeta`` is a command-line application that executes user queries and metadata retrieval from the NCBI suite of databases. The software is written in Python 3, using the ``BioPython`` module [@cock_biopython:_2009] to connect to, search, and download XML records with NCBI’s E-Utilities [@kans_entrez_2019]. The ``lxml`` package is utilised to perform XPath queries to retrieve nodes containing biological metadata of interest. ``SQLite`` is employed as the database management system for storing fetched records, as implemented with the ``sqlite3`` python module. Accessory scripts are provided to supply external annotation files, to join tables within the local database so as to re-create the relational database structure, and finally to export the database as tabular text for downstream analyses. ``NCBImeta`` currently interfaces with the molecular and literature databases described in Table \ref{NCBI_databases} [@noauthor_entrez_2016; @noauthor_ncbi_nodate].
+``NBCImeta`` is a command-line application that executes user queries and metadata retrieval from the NCBI suite of databases. The software is written in Python 3, using the ``BioPython`` module [@cock_biopython:_2009] to connect to, search, and download XML records with NCBI’s E-Utilities [@kans_entrez_2019]. The ``lxml`` package is utilised to perform XPath queries to retrieve nodes containing biological metadata of interest. ``SQLite`` is employed as the database management system for storing fetched records, as implemented with the ``sqlite3`` python module. Accessory scripts are provided to supply external annotation files, to join tables within the local database so as to re-create the relational database structure, and finally to export the database as tabular text for downstream analyses. ``NCBImeta`` currently interfaces with the molecular and literature databases described in Table \ref{NCBI_databases} [@noauthor_entrez_2016].
 
 : NCBI databases supported in NCBImeta. \label{NCBI_databases}
 
