+++ paper/paper.bib
@@ -220,13 +220,6 @@ Conclusions: A web server providing a convenient way of identifying acquired ant
 	file = {Eppinger et al. - 2014 - Genomic Epidemiology of the Haitian Cholera Outbre.pdf:files/787/Eppinger et al. - 2014 - Genomic Epidemiology of the Haitian Cholera Outbre.pdf:application/pdf}
 }
 
-@misc{noauthor_ncbi_nodate,
-	title = {{NCBI}},
-	url = {https://www.ncbi.nlm.nih.gov/},
-	urldate = {2019-10-16},
-	journal = {National Centre for Biotechnology Information}
-}
-
 @misc{wickham_ggplot2:_2016,
 	address = {New York},
 	title = {ggplot2: {Elegant} {Graphics} for {Data} {Analysis}},
