### 2020-01-07 #1990 @ktmeaton
I removed the redundant website reference NCBI (n.d.) which was flagged as MISSING DOIs. The information contained in Table 1 comes from Entrez Help (2016) which is a more substantial source.
