### 2020-01-22 a68d17a
Make output dir instead of raising error
