### 2020-01-22 c640928
Remove the output dir error class [skip ci]
