### 2020-01-22 619e1db
v0.6.2 bioconda and output dir
