+++ .travis.yml
@@ -3,7 +3,7 @@
 # Whitelist
 branches:
   only:
-    - master
+    - dev
 
 language: python    # this works for Linux but is an error on macOS or Windows
 jobs:
