### 2020-01-22 f820068
bioconda installation and media as raw links [skip ci]
