+++ .travis.yml
@@ -1,60 +1,48 @@
-# Travis-CI Configuration File
-
-# Whitelist
 branches:
   only:
-    - master
-
-language: python    # this works for Linux but is an error on macOS or Windows
+  - master
+language: python
 jobs:
   include:
-#-----------------------------------------------------------------------------#
-#                                    MACOS                                    #
-#-----------------------------------------------------------------------------#
-# macOS xcode9.4 = Python 3.6.5
-# macOS xcode10.2 = Python 3.7.3
-# macOS xcode11.2 = Python 3.7.4?
-    - name: "Python 3.7.4 on macOS"
-      os: osx
-      osx_image: xcode11.2  # Python 3.7.4 running on macOS 10.14.4
-      language: shell       # 'language: python' is an error on Travis CI macOS
-
-#-----------------------------------------------------------------------------#
-#                                    LINUX                                    #
-#-----------------------------------------------------------------------------#
-
-    - name: "Python 3.5 on Xenial Linux"
-      python: 3.5
-    - name: "Python 3.6 on Xenial Linux"
-      python: 3.6
-    - name: "Python 3.7.4 on Xenial Linux"
-      python: 3.7.4        # this works for Linux but is ignored on macOS or Windows
-    - name: "Python 3.8 on Xenial Linux"
-      python: 3.8
-
-
+  - name: Python 3.7.4 on macOS
+    os: osx
+    osx_image: xcode11.2
+    language: shell
+  - name: Python 3.5 on Xenial Linux
+    python: 3.5
+  - name: Python 3.6 on Xenial Linux
+    python: 3.6
+  - name: Python 3.7.4 on Xenial Linux
+    python: 3.7.4
+  - name: Python 3.8 on Xenial Linux
+    python: 3.8
 before_install:
-  - python3 --version
-
-# command to install dependencies
+- python3 --version
 install:
-  - pip3 install --upgrade pip       # all three OSes agree about 'pip3'
-  - pip3 install . -r requirements.txt
-  - pip3 install codecov==2.0.15 pytest==5.3.1 coverage==4.5.4 pytest-cov==2.8.1
-
-
-# command to run tests
+- pip3 install --upgrade pip
+- pip3 install . -r requirements.txt
+- pip3 install codecov==2.0.15 pytest==5.3.1 coverage==4.5.4 pytest-cov==2.8.1
 script:
-  - python3 -m coverage run -m pytest --cov=ncbimeta --cov-report=xml test/test_errors.py test/test_utilities.py test/test_ncbimeta.py test/test_annotateconcatenate.py test/test_annotatereplace.py test/test_join.py test/test_export.py
-  # Fetch records
-  - NCBImeta.py --flat --config example/config.yaml
-  # Database Annotation
-  - NCBImetaAnnotateReplace.py --database example/yersinia_pestis_db.sqlite --annotfile example/annot.txt --table BioSample
-  - NCBImetaAnnotateConcatenate.py --database example/yersinia_pestis_db.sqlite --annotfile example/annot.txt --table BioSample
-  # Database table join
[…]
