### 2020-01-24 cb49712
Document the bioconda repo master branch switch
