### 2020-01-24 8f0c064
more pwd checks
