+++ .travis.yml
@@ -43,8 +43,8 @@ after_success:
     git clone https://github.com/ktmeaton/bioconda-recipes.git;
     cd bioconda-recipes;
     git checkout master;
-    git user.name = $GITHUB_USER;
-    git user.email = $GITHUB_USER_EMAIL;
+    git config --global user.name = $GITHUB_USER;
+    git config --global user.email = $GITHUB_USER_EMAIL;
     git remote add upstream https://github.com/bioconda/bioconda-recipes.git;
     git remote rm origin;
     git remote add origin https://${GITHUB_USER}:${GITHUB_API_KEY}@github.com/${TRAVIS_REPO_SLUG}.git > /dev/null 2>&1;
