### 2020-01-24 a0fa57e
show git config
