+++ .travis.yml
@@ -40,6 +40,7 @@ after_success:
 #- bash <(curl -s https://codecov.io/bash)
 - if [[ "$TRAVIS_PYTHON_VERSION" == "3.5" ]]; then
     pwd;
+    git config -l;
     cd ../ ;
     pwd;
     git clone https://github.com/ktmeaton/bioconda-recipes.git;
@@ -52,6 +53,7 @@ after_success:
     git remote rm origin;
     git remote add origin https://${GITHUB_USER}:${GITHUB_API_KEY}@github.com/${TRAVIS_REPO_SLUG}.git > /dev/null 2>&1;
     git pull upstream master;
+    git config -l;
     git push -f -q https://${GITHUB_USER}:${GITHUB_API_KEY}@github.com/${TRAVIS_REPO_SLUG} master;
     cd ../NCBImeta/;
     pwd;
