### 2020-01-24 4554da9
fix incorrect slug usage
