+++ .travis.yml
@@ -46,15 +46,11 @@ after_success:
     git clone https://github.com/ktmeaton/bioconda-recipes.git;
     cd bioconda-recipes/;
     pwd;
-    # Make sure we're on the master branch
     git checkout master;
-    # Set essential parameters
     git config user.name $GITHUB_USER;
     git config user.email $GITHUB_USER_EMAIL;
-    # Retrieve changes from the upstream/original repo
     git remote add upstream https://github.com/bioconda/bioconda-recipes.git;
     git pull upstream master;
-    # Remove the original origin, change it to include the personal access token
     git remote rm origin;
     git remote add origin https://${GITHUB_USER}:${GITHUB_API_KEY}@github.com/${GITHUB_USER}/bioconda-recipes.git > /dev/null 2>&1;
     git config -l;
