### 2020-01-24 e48c1e6
remove dir checks
