+++ .travis.yml
@@ -39,13 +39,9 @@ script:
 after_success:
 #- bash <(curl -s https://codecov.io/bash)
 - if [[ "$TRAVIS_PYTHON_VERSION" == "3.5" ]]; then
-    pwd;
-    git config -l;
     cd ../ ;
-    pwd;
     git clone https://github.com/ktmeaton/bioconda-recipes.git;
     cd bioconda-recipes/;
-    pwd;
     git checkout master;
     git config user.name $GITHUB_USER;
     git config user.email $GITHUB_USER_EMAIL;
@@ -53,10 +49,8 @@ after_success:
     git pull upstream master;
     git remote rm origin;
     git remote add origin https://${GITHUB_USER}:${GITHUB_API_KEY}@github.com/${GITHUB_USER}/bioconda-recipes.git > /dev/null 2>&1;
-    git config -l;
     git push -f -q https://${GITHUB_USER}:${GITHUB_API_KEY}@github.com/${GITHUB_USER}/bioconda-recipes master;
     cd ../NCBImeta/;
-    pwd;
   fi
 deploy:
   provider: pypi
