### 2020-01-24 d8fc06b
conda var first pass
