+++ .travis.yml
@@ -39,6 +39,10 @@ script:
 after_success:
 #- bash <(curl -s https://codecov.io/bash)
 - if [[ "$TRAVIS_PYTHON_VERSION" == "3.5" ]]; then
+    program_name=ncbimeta;
+    bioconda_recipe=meta.yaml;
+    program_ver=`NCBImeta.py --version | cut -d " " -f2 | sed 's/v//g'`;
+    program_url=https://github.com/${TRAVIS_REPO_SLUG}/archive/v${program_ver}.tar.gz;
     cd ../ ;
     git clone https://github.com/${GITHUB_USER}/bioconda-recipes.git;
     cd bioconda-recipes/;
@@ -50,6 +54,19 @@ after_success:
     git remote rm origin;
     git remote add origin https://${GITHUB_USER}:${GITHUB_API_KEY}@github.com/${GITHUB_USER}/bioconda-recipes.git > /dev/null 2>&1;
     git push -f -q https://${GITHUB_USER}:${GITHUB_API_KEY}@github.com/${GITHUB_USER}/bioconda-recipes master;
+    cd recipes/;
+    dest_ver=`echo "{% set version = \"${program_ver}"\" %}`;
+    src_ver=`grep "set version" recipes/${program_name}/${bioconda_recipe}`;
+    dest_sha256=`wget -O- $URL | shasum -a 256 | cut -d " " -f 1`;
+    src_sha256=`grep "sha256" ${program_name}/${bioconda_recipe} | sed 's/^ *//g'`;
+    echo $program_name;
+    echo $bioconda_recipe;
+    echo $program_ver;
+    echo $program_url;
+    echo $dest_ver;
+    echo $src_ver;
+    echo $dest_sha256;
+    echo $src_sha256;
     cd ../NCBImeta/;
   fi
 deploy:
