+++ .travis.yml
@@ -38,7 +38,7 @@ script:
 #- NCBImetaExport.py --database example/yersinia_pestis_db.sqlite --outputdir example/
 after_success:
 #- bash < (curl -s https://codecov.io/bash)
-- if [[ "$TRAVIS_PYTHON_VERSION" == "3.5" ]]; then
+- if [[ "$TRAVIS_PYTHON_VERSION" == "3.5" && $TRAVIS_TAG ]]; then
     program_name=ncbimeta;
     bioconda_recipe=meta.yaml;
     program_ver=`NCBImeta.py --version | cut -d " " -f2 | sed 's/v//g'`;
@@ -60,9 +60,9 @@ after_success:
     dest_sha256=`wget -O- ${program_url} | shasum -a 256 | cut -d " " -f 1`;
     src_sha256=`grep "sha256" ${program_name}/${bioconda_recipe} | sed 's/^ *//g' | cut -d " " -f 2`;
     git checkout -b ${program_name}-${program_ver}
-    head ${program_name}/{bioconda_recipe};
-    sed -i "s/$src_ver/$dest_ver/g" ${program_name}/{bioconda_recipe} | sed "s/$src_sha256/$dest_sha256/g"
-    head ${program_name}/{bioconda_recipe};
+    head ${program_name}/${bioconda_recipe};
+    sed -i "s/$src_ver/$dest_ver/g" ${program_name}/${bioconda_recipe} | sed "s/$src_sha256/$dest_sha256/g"
+    head ${program_name}/${bioconda_recipe};
     cd ../../NCBImeta/;
   fi
 deploy:
