### 2020-01-24 98609ac
split up branch creation and check out
