### 2020-01-24 4053814
move conda update commands to script
