### 2020-01-24 2d5dda8
if statement spacing
