### 2020-01-24 c00f478
remove unnecessary tag condition
