+++ .travis.yml
@@ -49,5 +49,5 @@ deploy:
   skip_existing: true
   on:
     branch: master
-    condition: $TRAVIS_OS_NAME = linux && $TRAVIS_PYTHON_VERSION = "3.5" && $TRAVIS_TAG = v*
+    condition: $TRAVIS_OS_NAME = linux && $TRAVIS_PYTHON_VERSION = 3.5
     tags: true
