### 2020-01-24 41ed42e
switch zenodo citation to the all ver DOI [skip ci]
