+++ ncbimeta/NCBImetaAnnotateReplace.py
@@ -50,7 +50,7 @@ mandatory.add_argument('--annotfile',
 
 parser.add_argument('--version',
                     action='version',
-                    version='%(prog)s v0.6.2')
+                    version='%(prog)s v0.6.3')
 
 
 args = vars(parser.parse_args())
