+++ CHANGELOG.md
@@ -6,14 +6,20 @@ and this project "attempts" to adhere to [Semantic Versioning](http://semver.org
 
 ## [Development]
 
+
+## [v0.6.3] - 2020-0124 - Automating the Chain
+
 ### Added
 - conda_update.sh to automate conda recipe updating after Travis-CI tag run
+- ver_update.sh to automate version number updating of executables and build scripts
 - git update-index --add --chmod=+x conda_update.sh
+- git update-index --add --chmod=+x ver_update.sh
 
 ### Changed
 - 2020-01-23: the master branch switched to the bioconda-recipes repo. I'm uncertain about why this happened but I
 copied the master branch to a new branch called 'bc-tbd' (bioconda to-be-determined). To recover, I'm copying v0.6.2 to
 a new master branch.
+- 2020-01-24: Solved. This happened during Travis-CI testing. While working in the bioconda-recipes repo, my remote.origin.url was pointed towards the NCBImeta repo because of incorrected usage of the TRAVIS_REPO_SLUG env var. And I force pushed the bioconda commits onto the NCBImeta master branch. This can be undone with git reset --hard <commit> and then git push --force. Deleted branch 'bc-tbd'.
 
 ## [v0.6.2] - 2020-0123 - JOSS Review and Bioconda
 
