### 2020-01-24 a76ff01
use the bioconda recommended install code [skip ci]
