+++ .travis.yml
@@ -25,19 +25,19 @@ install:
 - pip3 install codecov==2.0.15 pytest==5.3.1 coverage==4.5.4 pytest-cov==2.8.1
 script:
 - python3 -m coverage run -m pytest --cov=ncbimeta --cov-report=xml test/test_errors.py
-#  test/test_utilities.py test/test_ncbimeta.py test/test_annotateconcatenate.py test/test_annotatereplace.py
-#  test/test_join.py test/test_export.py
-#- NCBImeta.py --flat --config example/config.yaml
-#- NCBImetaAnnotateReplace.py --database example/yersinia_pestis_db.sqlite --annotfile
-#  example/annot.txt --table BioSample
-#- NCBImetaAnnotateConcatenate.py --database example/yersinia_pestis_db.sqlite --annotfile
-#  example/annot.txt --table BioSample
-#- NCBImetaJoin.py --database example/yersinia_pestis_db.sqlite --anchor BioSample
-#  --accessory "BioProject Assembly SRA Nucleotide" --final Master --unique "BioSampleAccession
-#  BioSampleAccessionSecondary BioSampleBioProjectAccession"
-#- NCBImetaExport.py --database example/yersinia_pestis_db.sqlite --outputdir example/
+  test/test_utilities.py test/test_ncbimeta.py test/test_annotateconcatenate.py test/test_annotatereplace.py
+  test/test_join.py test/test_export.py
+- NCBImeta.py --flat --config example/config.yaml
+- NCBImetaAnnotateReplace.py --database example/yersinia_pestis_db.sqlite --annotfile
+  example/annot.txt --table BioSample
+- NCBImetaAnnotateConcatenate.py --database example/yersinia_pestis_db.sqlite --annotfile
+  example/annot.txt --table BioSample
+- NCBImetaJoin.py --database example/yersinia_pestis_db.sqlite --anchor BioSample
+  --accessory "BioProject Assembly SRA Nucleotide" --final Master --unique "BioSampleAccession
+  BioSampleAccessionSecondary BioSampleBioProjectAccession"
+- NCBImetaExport.py --database example/yersinia_pestis_db.sqlite --outputdir example/
 after_success:
-#- bash < (curl -s https://codecov.io/bash)
+- bash < (curl -s https://codecov.io/bash)
 - if [[ $TRAVIS_PYTHON_VERSION == 3.5 && $TRAVIS_TAG == v* ]]; then
     ./conda_update.sh;
   fi
