### 2020-01-24 49e4bab
hidden script names and autobump script [skip ci]
