### 2020-01-24 4446f94
remove bioconda scripts
