+++ .travis.yml
@@ -38,9 +38,6 @@ script:
 - NCBImetaExport.py --database example/yersinia_pestis_db.sqlite --outputdir example/
 after_success:
 - bash < (curl -s https://codecov.io/bash)
-- if [[ $TRAVIS_PYTHON_VERSION == 3.5 && $TRAVIS_TAG == v* ]]; then
-    ./.conda_update.sh;
-  fi
 deploy:
   provider: pypi
   user: ${GITHUB_USER}
