+++ .travis.yml
@@ -37,7 +37,7 @@ script:
   BioSampleAccessionSecondary BioSampleBioProjectAccession"
 - NCBImetaExport.py --database example/yersinia_pestis_db.sqlite --outputdir example/
 after_success:
-- bash < (curl -s https://codecov.io/bash)
+- bash <(curl -s https://codecov.io/bash)
 deploy:
   provider: pypi
   user: ${GITHUB_USER}
