### 2020-01-24 e4c17c7
slack travis-ci integration
