### 2020-01-24 #1990 @ktmeaton
I’ve updated the repositories Issues and Pull Requests, thank you @druvus for creating the bioconda recipe! Apologies for the delay, I just needed some time to research and experiment with how conda integration will work for long-term support.
