### 2020-01-29 #1990 @lpantano
@openjournals/joss-eics, this is almost there, but can somebody double-check the size of the paper will be good and not too long. One I did in the past was too long and we needed to update the archive and the version at the very last step. Trying to avoid that. Thanks!
