### 2020-01-30 #1990 @Kevin-Mattheus-Moerman
@lpantano thanks for your comment. I think the paper looks fine. It is slightly long but it look likes it needs the length to treat the components asked for and there are no sections/paragraphs (as far as I could tell from having a glance at the paper) that should be removed (e.g. if they would double documentation content).
You said "almost there", are there further developments we are waiting for? Let me know if this is ready to be processed for publication. Thanks.
