### 2020-02-03 #1990 @ktmeaton
Thank you for checking on the length requirements of the paper!

I've updated the most recent Zenodo archive to make sure the title, authors, and ORCID match the paper. The final repo version after review is 0.6.3 and the DOI is 10.5281/zenodo.3627329
