### 2020-02-03 #1990 @lpantano
@openjournals/joss-eics, we have all set up on our side now, I think :) Let us know.
